+++ cli/cmd/capture.go
@@ -72,7 +72,7 @@ the captured data.`,
 		go func() {
 			<-sigChan
 			port.Write([]byte("SET_CLED_OFF\r"))
-			color.HiRed("\r\n\r\nCancelled.")
+			color.HiRed("\r\n\r\nCancelled by user via Ctrl+C")
 			os.Exit(0)
 		}()
 
@@ -82,20 +82,26 @@ the captured data.`,
 		// clear serial buffer
 		utils.ClearBuffer(port, scanner);
 
-		// get board info & turn on CLED
+		// get board info, turn on CLED and check for iwdg reset
 		boardInfo, _ := utils.BoardCheck(port, scanner);
 		if (boardInfo.CledIsUsedForErrors == true) {
 			port.Write([]byte("SET_CLED_ON\r"))
+		}
+		if (boardInfo.LastResetWasIWDG) {
+			print(utils.AdvisoryStringIWDG(boardInfo))
 		}
 
 		capturedSamples := make([]utils.Sample, 0, samples)
 
 		// time estimate
 		var totalSeconds float64 = float64(samples-1) * float64(interval) + (float64(samples) * float64(0.85))
-		print(fmt.Sprintf("\r\nCapturing %d samples, at a %s interval\r\nEstimated time until completetion: %s\r\nCapture should be completed at around %s\r\n\r\n", samples, utils.TimeString(int64(interval)), utils.TimeString(int64(totalSeconds)), time.Unix((time.Now().Unix() + int64(totalSeconds)), 0).Format("15:04:05")))
+		print(fmt.Sprintf("\r\nCapturing %d samples, at a %s interval\r\nEstimated time until completion: %s\r\nCapture should be completed at around %s\r\n\r\n", samples, utils.TimeString(int64(interval)), utils.TimeString(int64(totalSeconds)), time.Unix((time.Now().Unix() + int64(totalSeconds)), 0).Format("15:04:05")))
 
 		// new progress bar
 		bar := progressbar.New(samples)
+
+		var crcAdvisoryDisplayed bool = false;
+		
 		for i := range samples {
 			// read and put values in the struct
 			var sample utils.Sample
@@ -107,7 +113,26 @@ the captured data.`,
 				sample.Timestamp = time.Now().UTC().Format("2006-01-02T15:04:05Z")
 			}
 
-			_, value := utils.ReadTemperature(port, scanner);
+			_, value, readError := utils.ReadTemperature(port, scanner);
+
+			var errorString string;
+			timestamp := time.Now().UTC().Format("15:04:05")
+			if (readError != nil) {
+				if (readError.Error() == "CRC") {
+					errorString = color.HiRedString("CRC error, perhaps your sensor line is noisy?")
+				} else if (readError.Error() == "PARSE") {
+					errorString = color.HiRedString("Error parsing the temperature...")
+				}
+				if (!crcAdvisoryDisplayed) {
+					fmt.Print(utils.AdvisoryStringCRC(boardInfo));
+					crcAdvisoryDisplayed = true;
+				}
+			}
+
+			if (readError != nil) {
+				bar.Describe(fmt.Sprintf("[%s]: %s",timestamp, errorString))
+			}
+
 			sample.Value = value
 
 			sample.ValueInFahrenheit = utils.ConvertCelsiusToFahrenheit(value)
@@ -119,7 +144,6 @@ the captured data.`,
 			// advance the progress bar
 			bar.Add(1)
 			
-			// print(fmt.Sprintf("Capturing samples: %.0f%% [%d/%d]\r\n", (float64(i+1)/float64(samples))*100, i+1, samples))
 			if (i+1 < samples) {
 				time.Sleep(time.Duration(interval) * time.Second)
 			}