+++ cli/utils/types.go
@@ -28,4 +28,6 @@ type Sample struct {
 type BoardInfo struct {
 	CledIsUsedForErrors bool
 	FirmwareVersion string
+	LastResetWasIWDG bool
+	SerialNumber string
 }